backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
greenlet==3.1.1
idna==3.10
limits==4.4.1
packaging==24.2
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.10.1
python-dotenv==1.1.0
pydantic-settings==2.10.1
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
wrapt==1.17.2
python-multipart==0.0.20